introduce spice_utils
